report-mailer/requirements.txt
pandas
python-docx
PyMuPDF
openai
python-dotenv
